Fixes Issue #2 - Heroku now successfully creates a table and saves responses to it
Before, the table creation sql was stored in a script which was being piped into psql, which assumes that the database is local. This worked locally and on Travis because the db was local, but Heroku uses a remote db. By moving the sql into the Database and running it through psycopg2, the table is successfully created. (When I first had tried this way, it wasn't working because I wasn't committing the cursor)